cmooney-001: adds a console injection warning
This commit adds a warning for users that might be convinced by an
attacker to open up the javascript console in the browser and inject
malicious code into their environment.  The hopes are that we prevent
some users of the CDP portal from being fooled into compromising their
session.

Also, creates a new security class that can be used for other security
related features in the future.  Someday we may want to hoist this out
into its own package.