store all raw multicall values in a "math" reducer
move the conversion functions out of multicall into the old reducers, so that the new one can store raw values. this will come in useful when applying the math functions from the SDK.

the math reducer is positioned around the old reducers rather than as a sibling, so that it can calculate derived values and insert them into places that existing components expect to simplify or eliminate their data-fetching logic.

"math reducer" is not a very good name, i know; consider it a placeholder...